requirements.txt
# Sprint4RPA - Requirements Principais
# Sistema HP Anti-Pirataria

# ==============================================================================
# CORE - Python Standard (não precisa instalar, mas listado para referência)
# ==============================================================================
# json, os, sys, pathlib, datetime, subprocess, argparse, typing, re, time

# ==============================================================================
# WEB SCRAPING - Selenium
# ==============================================================================
selenium>=4.15.0
webdriver-manager>=4.0.0

# ==============================================================================
# DATA PROCESSING
# ==============================================================================
pandas>=2.0.0
numpy>=1.24.0

# ==============================================================================
# MACHINE LEARNING & AI
# ==============================================================================
scikit-learn>=1.3.0
openai>=1.3.0              # Para scripts 2 e 4 (extração First IA e NLP)

# ==============================================================================
# NLP - Natural Language Processing
# ==============================================================================
vaderSentiment>=3.3.2      # Análise de sentimento em reviews
nltk>=3.8.1                # Recursos adicionais de NLP (opcional)

# ==============================================================================
# IMAGE PROCESSING
# ==============================================================================
Pillow>=10.0.0             # Processamento de imagens
requests>=2.31.0           # Download de imagens

# ==============================================================================
# GRAPH ANALYSIS
# ==============================================================================
networkx>=3.0              # Análise de grafos e redes
python-louvain>=0.16       # Detecção de comunidades

# ==============================================================================
# WEB DASHBOARD
# ==============================================================================
streamlit>=1.28.0          # Dashboard interativo
plotly>=5.17.0             # Visualizações interativas

# ==============================================================================
# UTILITIES
# ==============================================================================
python-dotenv>=1.0.0       # Gerenciamento de variáveis de ambiente
tqdm>=4.66.0               # Barras de progresso (opcional)

 Território de produtividade

# ==============================================================================
# NOTE: Algumas dependências são opcionais dependendo do uso
# ==============================================================================
# - openai: Obrigatória apenas para scripts 2 e 4
# - selenium: Obrigatória apenas para extração de dados
# - streamlit: Obrigatória apenas para dashboard
# - networkx: Obrigatória apenas para análise de grafos (script 7)

# ==============================================================================
# INSTALAÇÃO RECOMENDADA
# ==============================================================================
# pip install -r requirements.txt
#
# Para instalar apenas componentes específicos:
# - Extração: pip install selenium webdriver-manager pandas requests
# - Pipeline: pip install pandas numpy scikit-learn python-dotenv
# - IA/NLP: pip install openai vaderSentiment
# - Dashboard: pip install streamlit plotly pandas
# - Grafos: pip install networkx python-louvain

